+++ flake8_docstrings_complete/__init__.py
@@ -55,6 +55,11 @@ RETURN_NOT_IN_DOCSTR_MSG = (
     f"{RETURN_NOT_IN_DOCSTR_CODE} function/ method that returns a value should have the returns "
     f"section in the docstring{MORE_INFO_BASE}{RETURN_NOT_IN_DOCSTR_CODE.lower()}"
 )
+RETURN_IN_DOCSTR_CODE = f"{ERROR_CODE_PREFIX}008"
+RETURN_IN_DOCSTR_MSG = (
+    f"{RETURN_IN_DOCSTR_CODE} function/ method that does not return a value should not have the "
+    f"returns section in the docstring{MORE_INFO_BASE}{RETURN_IN_DOCSTR_CODE.lower()}"
+)
 TEST_FILENAME_PATTERN_ARG_NAME = "--docstrings-complete-test-filename-pattern"
 TEST_FILENAME_PATTERN_DEFAULT = r"test_.*\.py"
 TEST_FUNCTION_PATTERN_ARG_NAME = "--docstrings-complete-test-function-pattern"